Update modules, fix reattach env vars.
Update to use the release versions of go-plugin and
terraform-plugin-test, instead of replacing them.

Run `go mod tidy` and drop a fairly shocking amount of entries from
go.mod and go.sum.

Update the comment in runProviderCommand about the environment variable
we generate, and set both a 0.12 and 0.13 version of the variable. This
is annoying, but not the end of the world. Unfortunately, with the way
0.12 and 0.13 handle providers without provider blocks, we get subtly
different namespaces, so we have to set both so they'll get picked up
appropriately.